+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.1
 - Allow autoconfig of filter via url-patterns property (fixes [#3](https://github.com/lukashinsch/spring-boot-actuator-user-agent-metrics/issues/3))
+- Cache parsing of user agent string (fixes [#6](https://github.com/lukashinsch/spring-boot-actuator-user-agent-metrics/issues/6))
 
 ## 0.1.0
 - First release